Chore: Update pre-commit dependencies to latest
ERROR! the role 'lf-dev-libs' was not found in <>
The error appears to be in '': line 5, column 7, but may
be elsewhere in the file depending on the exact syntax
problem.

Update role name to match the role path on GHA and
Jenkins jobs.

Add the GHA workflow for verify.
Update tox and pre-commit dependencies.

Ref: https://github.com/ansible/molecule/issues/1042
Change-Id: I95c541dfd3dce731e487051477cfaaafedbb155e
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>